requirements.txt
appdirs==1.4.3
coverage==4.4.1
Django==1.11.7
django-appconf==1.0.2
django-autoslug==1.9.3
django-compressor==2.2
django-debug-toolbar==1.9.1
django-haystack==2.6.1
django-ipware==1.1.6
django-pagedown==0.1.3
django-uuslug==1.1.8
jieba==0.39
jsonpickle==0.9.5
markdown2==2.3.5
mistune==0.8.1
olefile==0.44
packaging==16.8
Pillow==4.3.0
Pygments==2.2.0
#PyMySQL==0.7.11
pyparsing==2.2.0
python-memcached==1.58
python-slugify==1.2.4
pytz==2017.3
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
sqlparse==0.2.4
Unidecode==0.4.21
urllib3==1.22
webencodings==0.5.1
WeRoBot==1.1.1
Whoosh==2.7.4

certifi==2017.7.27.1
dj-database-url==0.4.2
dj-static==0.0.6
django-bootstrap3==9.1.0
gunicorn==19.7.1
static3==0.7.0
psycopg2>=2.6.1
xadmin>=0.6.1
